fix int/size_t type
git-svn-id: http://svn.code.sf.net/p/xmlrpc-c/code@532 98333e67-4a24-44d7-a75c-e53540dd3050

+++ trunk/xmlrpc-c/src/xmlrpc_server_abyss.c
@@ -157,13 +157,13 @@ getBody(xmlrpc_env *        const envP,
 
     if (trace)
         fprintf(stderr, "XML-RPC handler processing body.  "
-                "Content Size = %u bytes\n", contentSize);
+                "Content Size = %u bytes\n", (unsigned)contentSize);
 
     body = xmlrpc_mem_block_new(envP, 0);
     if (!envP->fault_occurred) {
-        unsigned int bytesRead;
+        size_t bytesRead;
         const char * chunkPtr;
-        int chunkLen;
+        size_t chunkLen;
 
         bytesRead = 0;
